requirements.txt
Flask
numpy
pandas
mecab-python3
gensim
google-cloud-storage
